src/requirements.txt
asgiref==3.2.7
Django==2.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
